@@ -1,6 +1,7 @@
 package com.kdao.light;
 
 import com.github.lkqm.spring.jpa.repository.config.EnableMybatisJpaRepositories;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
@@ -12,13 +13,14 @@ import org.springframework.cache.annotation.EnableCaching;
 @EnableCaching
 @SpringBootApplication
 @EnableMybatisJpaRepositories
+@Slf4j
 public class LightControllerApplication {
     public static void main(String[] args) {
         try {
             SpringApplication.run(LightControllerApplication.class, args);
         }catch (Exception e){
             //加入捕捉错误，启动不了的时候可以看到原因
-            e.printStackTrace();
+            log.error("启动失败",e);
         }
 
     }
